backend/app/requirements.txt
alembic==1.10.4
aws-cfn-bootstrap==2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
docker==6.0.1
docopt==0.6.2
docutils==0.14
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lockfile==0.11.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.0
pipreqs==0.4.13
PyJWT==2.6.0
pystache==0.5.4
python-daemon==2.2.3
requests==2.28.2
simplejson==3.2.0
six==1.16.0
SQLAlchemy==2.0.12
typing-extensions==4.5.0
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.2.3
yarg==0.1.9
zipp==3.15.0